+++ scrum-tools-ui/e2e/specs/errors.spec.ts
@@ -1,0 +1,11 @@
+import { expect } from "@playwright/test";
+import test from "../fixtures";
+
+test.describe("errors", () => {
+  test("general error", async ({ helper, page, locators }) => {
+    await page.goto("/error");
+    await expect(locators.errors.general.title).toBeVisible();
+
+    await helper.visualComparison();
+  });
+});